refactor(auth): simplify authentication middleware by moving logic to web middleware and exporting session cookie function
